requirements.txt
Django
djangorestframework
django-filter
django-redis
